Refactor pinch zoom and content offset calculation to view model

+++ Yomu/Screens/ChapterPageList/ZoomScale.swift
@@ -12,7 +12,7 @@ import UIKit
 /// A data structure that represents zoom scale
 struct ZoomScale: CustomStringConvertible {
   /// Scale in 1 based, 1 -> 100%
-  let scale: CGFloat
+  let scale: Double
 
   /// String representation of zoom scale,
   /// will automatically multiply the scale by 100
@@ -26,15 +26,15 @@ struct ZoomScale: CustomStringConvertible {
   ///    it's considered as 100 based scale (I believe no one wants to zoom in by 1000%)
   ///
   ///  - returns: zoom scale with base 1  (1 -> 100%)
-  static private func normalize(scale: CGFloat) -> CGFloat {
+  static private func normalize(scale: Double) -> Double {
     return scale > 10 ? (scale / 100) : scale
   }
 
-  init(scale: CGFloat) {
+  init(scale: Double) {
     self.scale = ZoomScale.normalize(scale: scale)
   }
 
   init(scale: String) {
-    self.init(scale: CGFloat(Double(scale)!))
+    self.init(scale: Double(scale)!)
   }
 }
